BugFix: update thermovariable dictionary correctly
The thermo variables in the multi-domain simulation are compiled to the end of the y vector from the end towards the head of the vector, which is not necessarily k-1 space away from its original position. To make sure the thermo variable dict is updated correctly, we need to update it while we are filling the thermo variable to their new position.